+++ packages/agent/src/certificate.ts
@@ -150,6 +150,7 @@ export interface CreateCertificateOptions {
 
 export class Certificate {
   public cert: Cert;
+  public readonly rawCert: ArrayBuffer;
 
   /**
    * Create a new instance of a certificate, automatically verifying it. Throws a
@@ -191,6 +192,7 @@ export class Certificate {
     // Default to 5 minutes
     private _maxAgeInMinutes: number = 5,
   ) {
+    this.rawCert = certificate;
     this.cert = cbor.decode(new Uint8Array(certificate));
   }